Modify client and server example
