+++ lib/async/job/generic.rb
@@ -4,41 +4,27 @@
 # Copyright, 2024, by Samuel Williams.
 
 module Async
-	module Job
-		class Generic
-      def initialize(id, *args, scheduled_at: nil, **kwargs)
+  module Job
+    class Generic
+      def self.enqueue(...)
+        self.new(...).enqueue
+      end
+
+      def initialize(id, scheduled_at: nil)
         @id = id
         @scheduled_at = scheduled_at
-        @args = args
-        @kwargs = kwargs
       end
 
       attr :id
       attr :scheduled_at
 
-      def [](key)
-        return @id if key == 'id'
-        return @scheduled_at if key == 'scheduled_at'
-      end
-
-      def to_json(*args)
-        {
-          id: @id,
-          args: @args,
-          kwargs: @kwargs
-        }.to_json
-      end
-
-      def perform(*args, **kwargs)
+      def serialize
         raise NotImplementedError
       end
 
-      def self.call(hash)
-        kwargs = hash['kwargs'].each_with_object({}) do |(k, v), memo|
-          memo[k.to_sym] = v
-        end
-        self.new(hash['id']).perform(*hash['args'], **kwargs)
+      def call
+        raise NotImplementedError
       end
-		end
-	end
+    end
+  end
 end
